Ignore mouse events where globalPos contains NaN
A NaN value in globalPos can cause an endless recursion in
QGuiApplicationPrivate::processMouseEvent() when e->enhancedMouseEvent()
is true and this is not a move event.

When an event contains a NaN in globalPos, print a warning and ignore
the event.

Fixes: QTBUG-86207
Change-Id: I8789e465921789569715e52c98193f91945ea982
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 63bbbdc4b8dba7be5690065dc6a4ffa5a85b1b32)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>